Remove unmaintained and inaccurate document.
There are some commits I love to make...  This is not one of them.  This
document was the work of many people, both the people who created it
originally, plus all the translators who have spent considerable time
making it more available in other languages.  And to all these people, I
apologize for the need to do this.  Unfortunately, it has not been
maintained, is not accurate, and no one has stepped forward to revamp
(and test!) the _whole thing_ to bring it up to date.  Unfortunately, a
little fix here and little fix there only makes things less wrong.  This
is not what OpenBSD docs are supposed to be.

Per discussion with JMC, Henning, Markus.