spotted by jasper.  also delete this file, as noted when doing the other purges:
the project has in the past received a small pittance of money for sales
referrals to amazon, but it is wrong to accept this bribe money from such
people.  instead, we will let people who want these books make their own
decisions as to where to purchase books.

for more details, read
http://www.cjr.org/the_news_frontier/why_amazon_caved_and_what_it_m.php?page=all