Remove some support sites that don't mention anything about OpenBSD on their website.
I got as far as Italy before google started wanting captchas for all
my site: searches so anyone wanting to pick this up, start after that point.

There's a comment at the top of support.html that we may do this without
notification so I'm not contacting site owners about this.