/var/account/acct is time_t dependent, so accounting is hosed when upgrading.
My proposed fix: rm /var/account/acct with the other rm'd files.  It
will be recreated (in the new format) with the next boot, which will
occur after you reinstall your packages, when your system is ready for
use.  For non-accounting users, this is a non-event and a no-change (the
file won't exist).  It appears the summary files, /var/account/usracct
and /var/account/savacct are not an issue.

If active users (and coders!) of accounting could verify this or maybe
come up with a better process, I'd appreciate it -- I'm not a user of
this feature.