As noticed by Jan Stary, the porter's guide said that OpenBSD doesn't have -lcrypt because DES crypt is in libc.
I think this is just meant to tell people that if some s/w is trying to
use -lcrypt it can just be removed from linker flags. Same applies to
librt / libdl so adjust the text to say "the functions provided by.."
and mention those libs too.

Retain some mention of DES in that "The <tt>crypt()</tt> function does
not support DES, only bcrypt."