each faq page has header and footer links that look like this:
[faq index] [to section x - some topic] [to section y - another topic]

that can be very useful if the faq you're reading only covers one topic,
or if the pages are meant to be read in order. it makes sense to have
them on the pf and smtpd faq pages, but the main faq contains a variety
of unrelated topics, so they're not actually very useful there. remove.

while here, make a few subject titles more concise.

ok tb