Use -r<branch> in "cvs up" commands for -stable branches (-r was used in checkout but not updates; specifying this makes it easier for users to move from e.g. 6.2-stable to 6.3-stable). Difference between anoncvs.html and faq5 pointed out by nawi at nawi.is.
Use -A for cvs up on -current updates, again to help people switching
branch.

Place -A and -r<branch> at the same place in command lines to give a cue to
users what they're doing, but don't go as far as a full description of how
sticky tags work here, it's not relevant to most readers of these docs and
is easy to find in cvs(1) given the flags.

discussion/ok tb@